### 2023-02-18 #5169 @editorialbot
[PRE REVIEW]: Quasi-Monte Carlo Methods in Python
**Submitting author:** @tupui (<a href="http://orcid.org/0000-0001-9816-1416">Pamphile ROY</a>)
**Repository:** https://github.com/tupui/scipy
**Branch with paper.md** (empty if default branch): joss_qmc
**Version:** 1.10
