### 2023-02-19 #5169 @arfon
@tupui – thanks for your submission to JOSS. Could you please link here the models/modules/components contributed to `scipy` here please?
