### 2023-02-19 #5169 @tupui
Thank you @arfon. Here are the files we contributed to SciPy:

https://github.com/tupui/scipy/blob/main/scipy/stats/qmc.py
https://github.com/tupui/scipy/blob/main/scipy/stats/_qmc.py
https://github.com/tupui/scipy/blob/main/scipy/stats/_sobol_direction_numbers.npz
https://github.com/tupui/scipy/blob/main/scipy/stats/_sobol.pyx
https://github.com/tupui/scipy/blob/main/scipy/stats/_sobol.pyi
