+++ joss_qmc/paper.bib
@@ -26,6 +26,7 @@
   year = {2020},
   archivePrefix = {arXiv},
   eprint = {2008.08051},
+  doi = {10.1007/978-3-030-98319-2_4}
 }
 
 @misc{scipy2021qmc,
