+++ scipy/stats/_multicomp.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+import numpy as np
+
+
+__all__ = [
+]
