### 2023-02-23 90e3710
ENH: accept variable imbalanced groups
