### 2023-02-23 4c172b5
MAINT/DOC: mypy and refactor the doc
