### 2023-02-24 e1d6d08
TST: add balanced design cases
