### 2023-02-24 5335d1e
ENH: add allowance
