### 2023-02-24 6c6b455
ENH: add confidence interval around the mean
