### 2023-03-01 de54854
Apply suggestions from code review
Co-authored-by: Matt Haberland <mhaberla@calpoly.edu>
