### 2023-03-02 4654005
ENH/TST: add some input validation
