### 2023-03-03 6d1134a
MAINT: change obs in tests for samples
