### 2023-03-03 f051837
TST: compare with ttest_ind for 1 sample
