### 2023-03-03 d4539cc
MAINT: cache previous value of CI
