### 2023-03-03 5e16a49
ENH: add alternative for CI
