### 2023-03-03 d4a82cb
BUG: fix CI shape for alternatives
