### 2023-03-03 df3f326
TST: add str test
