### 2023-03-04 cee941f
Update scipy/stats/_multicomp.py
Co-authored-by: Matt Haberland <mhaberla@calpoly.edu>
