### 2023-03-04 eff1092
Update scipy/stats/_multicomp.py
Co-authored-by: Matt Haberland <mhaberla@calpoly.edu>
