### 2023-03-04 #5169 @arfon
@tupui – apologies for the delay here. I'm going to ask the JOSS editorial team to weigh in on whether they think this passes the substantial scholarly effort test we have at JOSS.

For JOSS editors:

```
cloc scipy/scipy/stats/*qmc.py
       2 text files.
       2 unique files.
       0 files ignored.

Language                     files          blank        comment           code
Python                           2            501           1465            856
SUM:                             2            501           1465            856
```

```
cloc scipy/scipy/stats/_sobol*
       2 text files.
       2 unique files.
       1 file ignored.

Language                     files          blank        comment           code
Cython                           1             83            137            217
Python                           1              8              0             46
SUM:                             2             91            137            263
```
