### 2023-03-04 #5169 @tupui
Thank you for the update @arfon.

I hope you will find this scholarly sufficient.

The first PR introducing this new module in SciPy was one of the largest PR SciPy had ever seen (more than 600 comments.) Behind the scene, the most renown researchers in the QMC field participated, and as I said in the paper, this directly lead to the publication of at least one paper. Now that the code is released, we have seen a lot of interest from various large projects to take advantage of this module.
