### 2023-03-06 9a3a77a
BUG: fix array casting validation
