### 2023-03-27 4ebb47d
DOC: stats.logrank: refine notes to documentation and comments to code
