+++ scipy/stats/_survival.py
@@ -410,21 +410,20 @@ def logrank(
     Parameters
     ----------
     x, y : array_like or CensoredData
-        Samples to compare based on their survival functions.
+        Samples to compare based on their empirical survival functions.
     alternative : {'two-sided', 'less', 'greater'}, optional
         Defines the alternative hypothesis.
 
-        The null hypothesis is that the survival functions underlying
-        `x` and `y` are equal. The following alternative
+        The null hypothesis is that the survival functions of the distributions
+        underlying `x` and `y` are identical. The following alternative
         hypotheses are available (default is 'two-sided'):
 
         * 'two-sided': the survival functions underlying `x` and `y`
           are unequal.
-        * 'less': the survival functions underlying `x`
-          is less than the survival function underlying `y`.
-        * 'greater': the survival functions underlying `x`
-          is greater than the survival function underlying
-          `y`.
+        * 'less': the survival function underlying `x` tends to be less
+          than the survival function underlying `y`.
+        * 'greater': the survival function underlying `x` tends to be greater
+          than the survival function underlying `y`.
 
     Returns
     -------
@@ -432,7 +431,9 @@ def logrank(
         An object containing attributes:
 
         statistic : float ndarray
-            The computed statistic of the test.
+            The computed statistic (defined below). Its magnitude is the
+            square root of the magnitude returned by most other logrank test
+            implementations.
         pvalue : float ndarray
             The computed p-value of the test.
 
@@ -442,29 +443,39 @@ def logrank(
 
     Notes
     -----
-    The logrank test [1]_ compares the expected events that occured at
-    times with respect to expected events if both survival curves where the
-    same
+    The logrank test [1]_ compares the observed number of events to
+    the expected number of events under the null hypothesis that the two
+    samples were drawn from the same distribution. The statistic is
 
     .. math::
 
         Z_i = \frac{\sum_{j=1}^J(O_{i,j}-E_{i,j})}{\sqrt{\sum_{j=1}^J V_{i,j}}}
         \rightarrow \mathcal{N}(0,1)
-        \\
-        E_{i,j} = O_j \frac{N_{i,j}}{N_j}
+
+    where
+
+    .. math::
+
+        E_{i,j} = O_j \frac{N_{i,j}}{N_j},
         \qquad
         V_{i,j} = E_{i,j} \left(\frac{N_j-O_j}{N_j}\right)
-        \left(\frac{N_j-N_{i,j}}{N_j-1}\right)
+        \left(\frac{N_j-N_{i,j}}{N_j-1}\right),
 
-    with ``i`` the group-here either `x` or `y`-, ``j`` the number of unique
-    times an event occured, ``N`` the number of subjects at risk (of the event
-    occuring) and ``O`` are the number of observed events.
+    :math:`i` denotes the group (i.e. it may assume values :math:`x` or
+    :math:`y`, or it may be omitted to refer to the combined sample)
+    :math:`j` denotes the time (at which an event occured),
+    :math:`N` is the number of subjects at risk just before an event occured,
+    and :math:`O` is the observed number of events at that time.
 
-    The ``statistic`` which is returned by `logrank` is to be linked to a
-    normal distribution instead of a chi-distribution. i.e. z and not z**2.
-    z**2 is traditionally seen in other software and linked to the chi-squared
-    distribution. Using the version which converges to the standard normal
-    distribution allows to provide altenatives.
+    The ``statistic`` $Z_x$ returned by `logrank` is the (positive or negative)
[…]
