### 2023-03-27 ee23a74
Update scipy/stats/_survival.py
