### 2023-03-27 b7020dc
Update scipy/stats/_survival.py
