### 2023-03-28 #5169 @jbytecode
👋👋👋 Dear @yangyushi, @ptmerz    👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**Quasi-Monte Carlo Methods in Python**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5169).

Thank you in advance!
