### 2023-03-29 #5309 @editorialbot
[REVIEW]: Quasi-Monte Carlo Methods in Python
**Submitting author:** @tupui (<a href="http://orcid.org/0000-0001-9816-1416">Pamphile ROY</a>)
**Repository:** https://github.com/tupui/scipy
**Branch with paper.md** (empty if default branch): joss_qmc
**Version:** v1.10.0-joss-qmc
