### 2023-04-02 #5309 @yangyushi
Hi @tupui thanks for bring QMC to scipy! I am delighted to walk through your manuscript/code/documentation, as well as to learn something about QMC.

For the manuscript, I would suggest following improvements

- The "Summary" section in the current proof seems to be empty, maybe try re-organising the content to have a summary.
- It is mentioned implicitly that QMC is available from scipy 1.7.0. I think it would help the readers more if the version number was stated explicitly in the, er, summary section.
- It would be great if some references could be provided for this claim: "In many cases, a QMC sequence can be used as a drop-in replacement for a random number sequence, yet they are proven to provide faster convergence rates (both in theory and practice)."

I tested the code in the scipy documentation (freshly installed Python3.10.4 +  Scipy1.10.1) and almost everything worked as expected.

I had a little problem when working with the [`PoissonDisk`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.qmc.PoissonDisk.html#scipy.stats.qmc.PoissonDisk) engine. As a naive user I would expect

```python
engine.random(n)
```

to return an array with size `(n, d)`, according to the documentation of [`random`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.qmc.PoissonDisk.random.html#scipy.stats.qmc.PoissonDisk.random).

I am aware that this is not true according the docstring of [`_random`](https://github.com/scipy/scipy/blob/c1ed5ece8ffbf05356a22a8106affcd11bd3aee0/scipy/stats/_qmc.py#L1816). However, `_random` is an abstract method, and the users won't have access to its docstring without reading the source code. This may not be ideal.

Maybe the [related documentation](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.qmc.PoissonDisk.random.html#scipy.stats.qmc.PoissonDisk.random) can be slightly improved? Or maybe we can generate a warning if the PoissonDisk engine failed to generate `n` samples?

Overall, I think this submodule is well designed and could be very helpful for the related community. I would recommend publication when the "summary" section is finished.
