### 2023-04-04 b9c1f98
Add summary section and reference for QMC performance
