### 2023-04-04 #5309 @tupui
Hi @yangyushi, thank you for your review! I updated the paper based on your suggestions.

For the reference, I used the same reference from Owen2019 which is really a good reference/summary of the current state of the field while being "mathematically sound."

Regarding the documentation, I appreciate the comment. I will think about how to make it more prominent in the documentation that `n` is not guaranteed and make a PR (hence it will be visible in SciPy 1.11). cc @mdhaber

During the review of this feature, we did discuss a warning but in general we are very "conservative" as we try to limit the impact we have on users code. Raising a warning introduce some churn and it was decided that it was too much in this case. On the other hand, for Sobol' we did judge that it was necessary to raise a warning in case of an incorrect number of samples being generated.
