### 2023-04-15 #5309 @ptmerz
I will start this review by saying that I really enjoyed both the software and the accompanying paper. I think this is a great contribution which I have no doubt is of great use to the community. It made my life as a reviewer almost too easy 😄

The "General checks" section seems all good, with the following caveats that are due to the fact that we are reviewing a contribution to a (very popular) library rather than a standalone software:
* While the software is available in the listed repository (https://github.com/tupui/scipy), it is a fork of https://github.com/scipy/scipy, and the submitted software has been contributed upstream. Any prospective user should likely download scipy directly rather than the fork that is listed in the paper.
* The authors listed have significantly contributed to the effort, as can be seen from the respective PR and the commit list therein. It is difficult to judge in which capacity other people's work on the library have significantly contributed to the effort. I do however think that the author list looks appropriate.

I don't think the points above are a problem, but wanted to bring them to the editor's attention.

I have a few suggestions regarding the paper:
* For the introduction of QMC methods, you are using (Owen, 2019) as a reference on line 21. In this context, I would suggest to use a reference to a foundational publication of the field, if you can trace one, or then omit the reference. At least in my understanding, a reference like "Quasi-Monte Carlo (QMC) methods (Owen, 2019)" implies that Owen 2019 is the original source of the method. Note that I think that the reference is completely fine to support other claims, like you do on lines 27/28.
* Also for the Owen 2019 citation, I'd suggest to add a link to https://artowen.su.domains/mc/, or to https://artowen.su.domains/mc/qmcstuff.pdf, since it's unpublished and may exist in different, possibly non-authorized, versions.
* Personally, I like the fact that you are citing GitHub pull requests. I still wanted to flag this to @jbytecode, since I am not sure what the editorial guide lines are in this respect.
* Typesetting nit: On line 44, the period should be a comma, while the period should be on line 45.

Overall, I think the paper is well written and does a good job at providing the information required by JOSS.

The software itself is in great shape and a pleasure to use. I really appreciate that the authors have chosen to implement it in a widely used library, and to adhere to its style and well-known APIs. It makes adoption a breeze, and underlines their claim that this can indeed serve as a drop-in replacement for random number sequences under the right circumstances. The thorough review process that the software has gone through means that the presented submission is of very high quality, as is the accompanying documentation.

Overall, I'd like to thank the authors for an excellent submission, and would definitely recommend it for publication as soon as the little details in the paper are ironed out.
