### 2023-04-16 #5309 @jbytecode
@ptmerz - Thank you for starting your review. Indeed, it is a comprehensive and clean job. Thank you.

Since any web site or GitHub repository can be cited, I don't see any barriers to cite specific pull requests.

@tupui - Could you please consider updating the repo and the manuscript as our review suggested? Thank you in advance.
