+++ joss_qmc/paper.bib
@@ -10,13 +10,20 @@
   doi={10.1038/s41592-019-0686-2}
 }
 
+@book{niederreiter1992,
+author = {Niederreiter, Harald},
+title = {Random Number Generation and Quasi-Monte Carlo Methods},
+publisher = {Society for Industrial and Applied Mathematics},
+year = {1992},
+doi = {10.1137/1.9781611970081},
+}
+
+
 @misc{owen2019,
   author = {Owen, Art B.},
   title = {Monte Carlo Book: the Quasi-Monte Carlo parts.},
   year = {2019},
-  publisher = {},
-  journal = {},
-  url = {}
+  url = {https://artowen.su.domains/mc/}
 }
 
 @article{owen2020,
