### 2023-04-18 #5309 @tupui
@ptmerz thank you for the thorough review and your kind words 😃

As per your suggestions: I added another well known reference in place of the first one, added the URL and fixed the typo.

@jbytecode let me know if there is anything else I can do. Thanks!
