### 2023-04-18 #5309 @tupui
A note about authorship. What cannot be necessarily seen in the commit history itself are offline discussions and reviews.

Art Owen and I had an extensive discussion that made part of this work possible. There is a ref. to some of that in the paper (owen2020), he also contributed most of the technical documentation, provided some R code and helped me debug my Python translation to scramble Halton, Sobol' n-bits and also LHS strength.

Matt Haberland did a tremendous job in terms of review and API design on the original PR. After that, he has become the "goto" reviewer for QMC in SciPy. He made critical remarks on the API and helped me fix more than a few bugs 😅

Lastly, Maximilian Balandat and I wrote all the code. Maximilian himself wrote the initial Sobol' and the multivariate distributions bits.

Besides them, we had some extra interactions which I acknowledged in the last section. Some of which, like with Sergei Kucherenko, were more important than other, but still a magnitude less important than the main contributors listed here.
