### 2023-04-18 #5309 @ptmerz
@tupui, thank you for taking care of the small review items, and also for the additional explanations about authorship. And congratulations again for a great contribution.

@jbytecode, I have finished my review checklist, and, as mentioned above, think it's an excellent contribution and would recommend it for publication in JOSS.
