### 2023-04-19 #5309 @jbytecode
@tupui - I've  just a pull request that includes minor corrections in both manuscript and bibtex files. Could you please review the changes? Please merge them if you are agree with them. Thank you in advance.

[PR](https://github.com/tupui/scipy/pull/23)
