### 2023-04-20 #5309 @tupui
* [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. *tag created for v1.10.0*

@jbytecode thank you.

I did another pass and to the best of my abilities fixed the typos.

I tagged my branch with `v1.10.0-joss-qmc` (there is already an existing official `v1.10.0`). I archived all of that in Zenodo [10.5281/zenodo.7849706](https://zenodo.org/record/7849706). I made clear in the description that this specific version should not be used (and I also added this message as a warning if someone is still using this code. This follows what was previously done for another JOSS article we made for SciPy, see this record https://zenodo.org/record/6781029)
